feat: mark read, interup

+++ app/src/main/java/info/hermiths/chatapp/model/resp/History.kt
@@ -21,7 +21,7 @@ data class Content(
     @SerializedName("msgSeq") val msgSeq: Int,
     @SerializedName("msgBody") val msgBody: String,
     val clientMsgID: String,
-    val status: Long,
+    val status: Int,
     val createTime: Long,
     val sessionId: String,
 )
